requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pymongo>=4.6.0
motor>=3.3.0
sentence-transformers>=2.2.0
torch>=2.0.0
pillow>=10.0.0
python-multipart>=0.0.6
numpy>=1.24.0